+++ NEWS.md
@@ -1,3 +1,7 @@
+# flscuts v0.1.1
+
+* bug in iterative_LSI that did not return svd matrix if resolution vector was longer than 1
+* Fix export_seurat_to_10x matrix
 
 # flscuts v0.1.0